Move Key into extension
